Move : move test code #32
